--- backend/.streamlit/utils.py
@@ -1,2 +1,0 @@
-def calculate_something(data):
-    return data * 2
